// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const PostStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]


export const MediaType = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]